Extract background loops from main.py into dedicated modules
Moves 5 background loops, heartbeat tracking, crash guard, and lifecycle
helpers into src/services/core/loops/ — reducing main.py from 851 to ~190
lines of pure service wiring. Pure structural refactor, no behavior changes.

Fixes #271
